Migrated load account data test to jest
